Split logging into new file for the future
